requirements.txt
# Combined Bot requirements
python-telegram-bot==21.0.1
requests==2.31.0
pandas==2.2.3
numpy==1.26.4
python-binance==1.0.19
aiohttp==3.9.3
pytz==2023.3
APScheduler==3.10.4
python-dotenv==1.0.0
matplotlib==3.8.2
Pillow==10.1.0
tabulate==0.9.0
ccxt==4.3.5
ta==0.11.0
websockets==13.0.1
plotly==5.22.0
openai==1.35.3
redis==5.0.8
setuptools==75.1.0
wheel==0.44.0
aiohttp==3.9.3
APScheduler==3.10.4
ccxt==4.3.5
matplotlib==3.8.2
numpy>=1.21,<2
openai==1.35.3
pandas==2.2.3
Pillow==10.1.0
plotly==5.22.0
python-binance==1.0.19
python-dotenv==1.0.0
python-telegram-bot==21.0.1
pytz==2023.3
redis==5.0.8
requests==2.31.0
ta==0.11.0
tabulate==0.9.0
telegram
websockets==13.0.1
